+++ models/cart-discount/src/count-on-line-item-units/src/count-on-line-item-units-draft/types.ts
@@ -1,10 +1,5 @@
 import type { TBuilder } from '@commercetools-test-data/core';
-import { TCtpCountOnLineItemUnitsInput } from '@commercetools-test-data/graphql-types';
-
-export type TCountOnLineItemUnitsDraftGraphql = Omit<
-  TCtpCountOnLineItemUnitsInput,
-  'excludeCount'
->;
+import { TCountOnLineItemUnitsDraftGraphql } from '../../types';
 
 export type TCreateCountOnLineItemUnitsDraftBuilder<
   TModel extends TCountOnLineItemUnitsDraftGraphql,
